search por nome e preco

+++ api/src/main/java/br/com/digisystem/api/model/Produto.java
@@ -26,6 +26,6 @@ public class Produto {
 	private int id;
 	
 	private String nome;
-	private double preco;
+	private float preco;
 
 }
